Created messageboxcontroller
