+++ src/Altinn.Platform/Altinn.Platform.Storage/Storage.Interface/Storage.Interface.csproj
@@ -7,7 +7,6 @@
   </PropertyGroup>
 
   <PropertyGroup>
-
     <Version>1.1.9-alpha</Version>
     <Description>Altinn.Platform.Storage.Interface is a package for models, httpclient methods for altinn platform storage</Description>
     <PackageTags>altinn studio, altinn.platform, Altinn.Platform.Storage.Interface</PackageTags>
@@ -27,7 +26,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />''
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
   </ItemGroup>
 
   <ItemGroup>
